backend connected to atlas db
